### 2025-07-01 #8527 @editorialbot
[PRE REVIEW]: Histomics Label
**Submitting author:** @manthey (<a href="http://orcid.org/0000-0002-4580-8770">David Manthey</a>)
**Repository:** https://github.com/DigitalSlideArchive/histomics-label
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
