### 2025-10-28 #8527 @Kevin-Mattheus-Moerman
@manthey can you clarify if you seek publication of just this code base: https://github.com/DigitalSlideArchive/histomics-label or also: https://github.com/DigitalSlideArchive/superpixel-classification ?

Secondly, can you edit your paper title to not just have the project name but also a short descriptive sentence? For instance something like an improved version of this:
`histomics-label: A software tool for the interactive development of machine-learning classifiers for whole slide pathology images`

Thanks.
