### 2026-01-30 #9918 @juliaferraioli
@manthey I did notice that your LICENSE file is not the standard Apache 2.0 license. You may want to update that to the full text found at [apache.org](https://www.apache.org/licenses/LICENSE-2.0.txt).
