### 2026-02-23 #9918 @Mmasoud1
Hi @juliaferraioli , thank you for the reminder. I apologize for the delay, I have been traveling overseas, which limited my availability, but I will be able to complete the review in the coming week.  But, before proceeding, I would like to disclose a past research collaboration with one of the authors, Dr. Cooper. Our collaboration ended in May 2021, which is outside the JOSS four-year window for recent collaborations. I believe I can provide an impartial review and request guidance on whether this requires a waiver.  I will, of course, respect the journal’s decision should it be preferable to assign the review to another reviewer.  Thank you.
