### 2026-03-15 #9918 @arfon
:wave: folks, EiC of JOSS here. I'm stepping in here to help @juliaferraioli with her editorial duties.

@Mmasoud1 – as to your conflict disclosure, thank you for sharing this, and please proceed with your review.
