### 2026-04-21 #9918 @Mmasoud1
@arfon I have reviewed the paper and tested the software. Overall, this work addresses an important problem in digital pathology, namely, the labeling of whole-slide image (WSI) regions, and provides a practical solution using an active learning workflow to enhance tile classification while enabling collaboration and iterative model refinement.

From hands-on testing, I was able to successfully install and run the Histomics Label plugin within the Digital Slide Archive (DSA). The whole system requires a Docker-based setup. The full pipeline (superpixel generation, feature extraction, labeling, training, and guided refinement) worked end-to-end as described. The system correctly generated expected outputs (annotations, features, models), and model behavior improved with iterative labeling, demonstrating an active learning workflow.

In general, the installation via Docker is reproducible and manageable, and the integration with the DSA ecosystem is well designed and practical. However, **additional clarification in the documentation particularly for cross-platform users would be beneficial.**

### Suggestions for Improvement

- **State of the field**
  Adding a brief comparison with existing tools (e.g., HistomicsML etc) to better position the contribution.

- **Documentation (functionality details)**
  It is recommended to expand documentation, potentially via a **GitHub Wiki**, including pages such as:
  - Setup instructions
  - Histomics Label demo (screenshots or video)
    - Include a clear start-to-finish example using a public dataset (e.g., TCGA).
  - Contribution guide
    -  Add explicit contribution and support guidelines (e.g., `CONTRIBUTING.md`, issue templates).
  - System requirements
  - Any other related pages

  -  Additionally, provide more details on model options, parameters, and components (e.g., segmentation models and training behavior).

- **Version clarity (UI vs repository)**
  In the DSA interface, the Histomics Label plugin appears as version `0.1.0 (experimental)`, while the repository does not clearly reflect versioning history or alignment with this label. It would be helpful to clarify version consistency between the UI, repository, and paper to avoid confusion for users and reviewers.

- **Active learning behavior and model performance**
  During testing, I observed that distinguishing background ("bg") from tissue regions remained challenging for the classifier, even after multiple annotation iterations. While this may be expected depending on feature representation or model design and robust, it would be helpful for the documentation to:
  - Clarify expected model behavior and limitations
  - Suggest best practices for improving performance (e.g., number of annotations, feature tuning)

- **Extensibility (custom models/classifiers)**
  It is not clear how users/developers can integrate their own classifiers or extend the existing learning pipeline. Providing documentation or examples on how to plug in custom models (e.g., replacing or extending the default classifier) would significantly improve the flexibility and adoption of the tool for different research workflows.

- **References**
  Fix missing or unclear citations (e.g., TCGA reference on page 3, line 76).

- **Author contributions**
  Consider adding a [CRediT-style ](http://dx.doi.org/10.1087/20150211) author contributions section in the paper for clarity.

With the improvements above, I would support acceptance.
