### 2026-05-03 #9918 @cinnetcrash
# JOSS Review: histomics-label

**Repository:** https://github.com/DigitalSlideArchive/histomics-label
**Review Date:** 2026-05-03
**Software Version:** v1.0.0 (June 2025)

I was able to install and run the program with Docker. It's easy but there are some elements below.

---

**histomics-label** is a Girder 3 plugin that facilitates active learning
workflows for annotating whole slide images (WSIs) in digital pathology. It
enables users to label superpixel regions on histological slides and iteratively
improve machine learning models through human-in-the-loop feedback. The software
integrates with the Digital Slide Archive (DSA) ecosystem, including HistomicsUI
and HistomicsTK.

The software addresses a genuine need in computational pathology: reducing the
annotation burden for training deep learning models on WSIs through active
learning. However, in its current state, the submission does not meet the minimum
requirements for JOSS publication and requires major revisions before acceptance.

## Major Issues (Must Be Resolved Before Acceptance)

### 1. Critically Insufficient Test Suite

The test suite consists of a single import check:

```python
@pytest.mark.plugin('histomics_label')
def test_import(server):
    assert 'histomics_label' in loadedPlugins()
```

This is entirely insufficient for a JOSS submission. JOSS requires that
"software should be tested (unit, functional, or integration tests, with
automated CI where appropriate)." The existing CI pipeline (CircleCI) only
runs linters.

**Required action:** Implement a meaningful test suite covering core
functionality (superpixel generation, label assignment, active learning
triggering, confidence ranking). A minimum coverage of 50% is expected.
Issue #16 must be resolved.

---
### 2. Missing Community Guidelines

There is no `CONTRIBUTING.md` or `CODE_OF_CONDUCT.md`. JOSS expects projects
to "provide guidelines for third parties wishing to: Contribute to the
software, Report issues or problems with the software, Seek support."

**Required action:** Add `CONTRIBUTING.md`, `CODE_OF_CONDUCT.md`, and GitHub
issue/PR templates.

---

### 3. No Statement of Need or Related Work Comparison

The software's relationship to existing tools in digital pathology annotation
(e.g., QuPath, Cytomine, Label Studio, CVAT, Labelbox) is never discussed.
It is unclear what specific gap histomics-label fills that existing tools do not.

**Required action:** Add a "Statement of Need" section to `paper.md` and/or
`README.rst` with explicit comparison to at least 3–4 related tools.

## Recommendations

I recommend **major revision** before this submission can be accepted for
publication in JOSS. The software itself is functional and addresses a
legitimate need in computational pathology. However, in its current state it
fails to meet the minimum requirements for JOSS on the following grounds:

1. The test suite is insufficient to verify correctness of core functionality.
2. No formal statement of need or comparison with related tools is provided.

I encourage the authors to address the major issues listed above. I would be
happy to re-review and and approve the submission once these items have been resolved.
